Remove backends:
The gating of serving iPXE binaries had one main
role. This was to fail the PXE request so that
BIOS or UEFI could move on to the next boot device.
The challenge is that gating at this layer means using
the requesters IP address. This means we must know the IP
address. And means we need to query DHCP. This coupling makes
it very difficult to work in a bring your own DHCP model. And
bring your own DHCP a highly desired capability.

Signed-off-by: Jacob Weinstock <jakobweinstock@gmail.com>